+++ src/main/java/org/radarcns/android/data/DataHandler.java
@@ -17,6 +17,7 @@
 package org.radarcns.android.data;
 
 import org.radarcns.android.kafka.ServerStatusListener;
+import org.radarcns.kafka.ObservationKey;
 import org.radarcns.topic.AvroTopic;
 
 import java.util.Map;
@@ -30,18 +31,14 @@ public interface DataHandler<K, V> extends ServerStatusListener {
     /** Get all caches */
     Map<AvroTopic<K, ? extends V>, ? extends DataCache<K, ? extends V>> getCaches();
 
-    /** Get the cache for a given topic.
-     * @return DataCache or null if the topic is not found.
-     */
-    <W extends V> DataCache<K, W> getCache(AvroTopic<K, W> topic);
 
     /**
      * Add a measurement using given cache.
-     * @param cache cache to add measurement to.
+     * @param topic topic to add measurement to.
      * @param key key of the measurement
      * @param value value of the measurement.
      */
-    <W extends V> void addMeasurement(DataCache<K, W> cache, K key, W value);
+    <W extends V> void addMeasurement(AvroTopic<K, W> topic, K key, W value);
 
     /**
      * Set maximum number of bytes a single cache may contain.
